+++ packages/beardb/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.0.11
+
+### Patch Changes
+
+- Simplified removed nested promise stuff
+- Updated dependencies
+  - @asmartbear/simplified@1.0.11
+  - @asmartbear/smarttype@1.0.17
+
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
 ### [1.0.10](https://github.com/asmartbear/beardb/compare/v1.0.9...v1.0.10) (2025-11-27)